update the migration and model to use numeric instead of decimal and float

+++ backend/db/migrations/20241101222228-create-spot.js
@@ -41,11 +41,11 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.FLOAT(9, 6),
+        type: Sequelize.NUMERIC(9, 6),
         allowNull: false,
       },
       lng: {
-        type: Sequelize.FLOAT(9, 6),
+        type: Sequelize.NUMERIC(9, 6),
         allowNull: false
       },
       name: {
